Refactor tree variable names for clarity
